Start eliminating unnecessary static_casts

+++ opensyde_tool/src/com_import_export/C_CieDcfEdsImportNodeSelectWidget.cpp
@@ -170,8 +170,7 @@ void C_CieDcfEdsImportNodeSelectWidget::keyPressEvent(QKeyEvent * const opc_KeyE
    bool q_CallOrg = true;
 
    //Handle all enter key cases manually
-   if ((opc_KeyEvent->key() == static_cast<int32_t>(Qt::Key_Enter)) ||
-       (opc_KeyEvent->key() == static_cast<int32_t>(Qt::Key_Return)))
+   if ((opc_KeyEvent->key() == Qt::Key_Enter) || (opc_KeyEvent->key() == Qt::Key_Return))
    {
       if (((opc_KeyEvent->modifiers().testFlag(Qt::ControlModifier) == true) &&
            (opc_KeyEvent->modifiers().testFlag(Qt::AltModifier) == false)) &&
