+++ opensyde_tool/src/system_views/system_update/update_package/C_SyvUpPacParamSetFileAddPopUp.cpp
@@ -146,8 +146,7 @@ void C_SyvUpPacParamSetFileAddPopUp::keyPressEvent(QKeyEvent * const opc_KeyEven
    bool q_CallOrg = true;
 
    //Handle all enter key cases manually
-   if ((opc_KeyEvent->key() == static_cast<int32_t>(Qt::Key_Enter)) ||
-       (opc_KeyEvent->key() == static_cast<int32_t>(Qt::Key_Return)))
+   if ((opc_KeyEvent->key() == Qt::Key_Enter) || (opc_KeyEvent->key() == Qt::Key_Return))
    {
       if (((opc_KeyEvent->modifiers().testFlag(Qt::ControlModifier) == true) &&
            (opc_KeyEvent->modifiers().testFlag(Qt::AltModifier) == false)) &&
